Move search settings loading code and routes sending to index.tsx This ensures it's done and single-spa load time not mount time And allows scigateway to know the plugin routes which it should mount the plugin on
