+++ packages/datagateway-search/src/settings.js
@@ -1,0 +1,4 @@
+export let settings;
+export const setSettings = (newSettings) => {
+  settings = newSettings;
+};
